requirements.txt
Flask
ultralytics
opencv-python
Pillow
